Add `retry_coordinator`, `user_agent` config
`retry_coordinator` is meant to be used by clients to e.g. update the
token when it expires, or to instruct aptible-api to retry the request
at the application layer (connection layer issues will be retried by
Faraday). This might be useful if we're hitting a Bad Gateway issue that
we'd like to retry (since we know that in this case, the API didn't
actually *see* the request).

`user_agent` is meant to let the client advertise who they are. This
will presumably be useful for e.g. the CLI to advertise its version, or
Sweetness (this could also be made more specific, including the
Sweetness stack, etc.).

This also tries to address several gravely misguided architectural
decisions in HyperResource, like caching connections in a thread pool
that's never cleaned (!), with keys that are so specific that they're
unlikely to be ever be reused (essentially making the thread pool
useless), or updating attributes on an object with a response that might
turn out to be invalid.