Retry errors on safe methods by default
This should minimize occasional errors we get making requests to our
APIs from various clients (but not completely eliminate them, since POST
requests will not be retried).

The retry schedule is intentionally short (we'll make at most 4 requests
over ~4 seconds) to avoid making tripping API users that don't normally
expect a request to take 4 seconds, but the `DefaultRetryCoordinator` is
designed to be easy to subclass in order to provide a different retry
schedule.

Note that a more bulletproof approach that would support POSTs as well
would be to use a client-generated token to identify unique requests,
like [EC2][0] and [Stripe][1] do, but obviously, that requires
server-side support (and even in that case, we *probably* wouldn't want
to support idempotency tokens for any method other than POST).

Let's see how far along this takes us. We can always contemplate
idempotency tokens afterwards if we still hit more than errors than
we're comfortable with.

[0]: http://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html
[1]: https://stripe.com/docs/api?lang=curl#idempotent_requests