Overhaul retries
This delegates all retry logic to our retry coordinator, which ensures
e.g. timeouts aren't handled by Faraday as a middleman.

I'm also adding a `without_retry` method that can be used to disable
retries altogether. This isn't a very clean approach, but unfortunately
considering we don't have a concept of API client instances (i.e.  our
only configuration is global), that's the best I can do right now.

This will be useful for cases where retries are explicitly not
desirable, which is specifically the case when Sweetness attempts

This does change the API exposed by retry coordinators, but as far as I
can tell this isn't a feature we ever publicized, and the only consumer
is Sweetness, which pins to a version of the library. That being said,
we'll still want to update the minor to reflect the breaking change
(we're < 1.0 so that's semver-compliant).