each_X: return enumerator when no block is passed
This allows callers to capture an enumerator for later reuse. More
broadly speaking, this is arguably the expected behavior in a Ruby
library when calling an enumerator method without a block.

I've also added a `#each_X` method for `embeds_many`, so that callers do
not need to differentiate between `embeds_many` and `has_many`
relationships when calling `each_X`.