Relax Faraday dependency
There appear to have only been minor feature additions and bug fixes in
Faraday between 0.9.2 and the current version (0.13.1), and we no longer
depend on Faraday internals (or much of Faraday) as of #33.

There's no particular urgency in upgrading ourselves, but pinning to
v0.9.x prevents customers from installing `aptible-resource` alongside
other gems that dependent on a more recent version of Faraday.