Fix HyperResource#method_missing block handling
HyperResource does not pass through the block in `method_missing` for a
link. As a result, this type of call fails (it ignores the blocks and
returns an `Enumerator`):

```
api = Aptible::Api::Resource.new(token: token)
acct = api.accounts.find { |a| a.handle == 'foo' }
```

We don't normally call this too often since we have our `has_many`
methods defined on the resources we query, but the above example is
actually a useful use case, which should be supported.

I've also removed a few exception-silencing statements in the process.